+++ src/Components/Monu/Loading/Loading.jsx
@@ -72,22 +72,7 @@ function Loading() {
 </Box>
  
     </div>
-    <div style={{display:"flex",alignItems:"center",justifyContent:"center"}}> 
-     {/* <Box padding='6' boxShadow='lg' bg='white'>
-        <div class="card">
-        <div class="shimmerBG media"></div>
-        <div class="p-32">
-          <div class="shimmerBG title-line"></div>
-          <div class="shimmerBG title-line end"></div>
-
-          <div class="shimmerBG content-line m-t-24"></div>
-          <div class="shimmerBG content-line"></div>
-          <div class="shimmerBG content-line"></div>
-          <div class="shimmerBG content-line"></div>
-          <div class="shimmerBG content-line end"></div>
-        </div>
-      </div>
-</Box> */}
+    <div style={{display:"flex",alignItems:"center",justifyContent:"center"}}>
  <Box padding='6' boxShadow='lg' bg='white'>
         <div class="card">
         <div class="shimmerBG media"></div>
@@ -134,7 +119,21 @@ function Loading() {
         </div>
       </div>
 </Box>
- 
+  <Box padding='6' boxShadow='lg' bg='white'>
+        <div class="card">
+        <div class="shimmerBG media"></div>
+        <div class="p-32">
+          <div class="shimmerBG title-line"></div>
+          <div class="shimmerBG title-line end"></div>
+
+          <div class="shimmerBG content-line m-t-24"></div>
+          <div class="shimmerBG content-line"></div>
+          <div class="shimmerBG content-line"></div>
+          <div class="shimmerBG content-line"></div>
+          <div class="shimmerBG content-line end"></div>
+        </div>
+      </div>
+</Box>
     </div>
 
     </div>
